Se realiza la implementación correspondiente de los metodos, clases y widgets requeridos para obtener las categorias con sus subcategorias desde el backend.

+++ lib/models/address.dart
@@ -1,7 +1,7 @@
 import 'dart:ffi';
 
 class Adress {
-  final Long id;
+  final int id;
   final String address;
   final int doorNumber;
   final String location;
